Mapped operations in LillyDAP class to _lillydap, with limitations  - using a singleton instance (LillyDAP *) for now (TODO: bind to self)  - not dealt with reference counting yet  - code does repeat itself (but that is not a big problem)  - at the bottom of the parse stack, we must still link to Python  - intermediate code needed for all callbacks to try and use a method

+++ include/lillydap/api.h
@@ -167,7 +167,7 @@ int lillyput_ldapmessage (LDAP *lil,
 int lillyput_dercursor (LillyDAP *lil, LillyPool qpool, dercursor dermsg);
 void lillyput_enqueue (LillyDAP *lil, struct LillySend *addend);
 bool lillyput_cansend (LillyDAP *lil);
-int lillyput_event (LDAP *lil);
+ssize_t lillyput_event (LDAP *lil);
 
 
 /* A parallel to ldap_open, filling the basic structure and returning
